+++ src/views/components/l13-diff-input/l13-diff-input.component.ts
@@ -1,12 +1,14 @@
 //	Imports ____________________________________________________________________
 
 import { L13Component, L13Element, L13Query } from '../../@l13/core';
+
+import { L13DiffListComponent } from '../l13-diff-list/l13-diff-list.component';
 import { L13DiffMenuComponent } from '../l13-diff-menu/l13-diff-menu.component';
 
 import { L13DiffInputViewModelService } from './l13-diff-input.service';
 import { L13DiffInputViewModel } from './l13-diff-input.viewmodel';
 
-import { parseIcons, setLabelText } from '../common';
+import { changePlatform, parseIcons, setLabelText } from '../common';
 import styles from '../styles';
 import templates from '../templates';
 
@@ -36,6 +38,8 @@ export class L13DiffInputComponent extends L13Element<L13DiffInputViewModel> {
 	
 	public menu:L13DiffMenuComponent;
 	
+	public list:L13DiffListComponent;
+	
 	public constructor () {
 		
 		super();
@@ -62,16 +66,21 @@ export class L13DiffInputComponent extends L13Element<L13DiffInputViewModel> {
 			
 		});
 		
-		this.input.addEventListener('keydown', (event) => {
+		this.input.addEventListener('keydown', ({ key, metaKey, ctrlKey, altKey, shiftKey }) => {
 			
 			const menu = this.menu;
 			
-			switch (event.key) {
+			switch (key) {
+				case 'F12': // Debug Mode
+					if (metaKey && ctrlKey && altKey && shiftKey) changePlatform();
+					break;
 				case 'Enter':
-					if (menu && menu.parentNode) {
-						const value = menu.getSelection();
-						if (value) this.input.value = value;
-						menu.remove();
+					if (menu) {
+						if (menu.parentNode) {
+							const value = menu.getSelection();
+							if (value) this.viewmodel.value = value;
+							menu.remove();
+						} else this.list.viewmodel.compare();
 					}
 					break;
 				case 'Tab':