+++ src/views/components/l13-diff-list/l13-diff-list.component.ts
@@ -8,13 +8,16 @@ import { L13DiffListViewModel } from './l13-diff-list.viewmodel';
 
 import { L13DiffActionsViewModelService } from '../l13-diff-actions/l13-diff-actions.service';
 
-import { isMacOs ,isMetaKey, isOtherPlatform, isWindows, parseIcons, removeChildren, vscode } from '../common';
+import { changePlatform, isMacOs, isMetaKey, isOtherPlatform, isWindows, parseIcons, removeChildren, scrollElementIntoView, vscode } from '../common';
 import styles from '../styles';
 import templates from '../templates';
 
 //	Variables __________________________________________________________________
 
 const actionsService = new L13DiffActionsViewModelService();
+
+enum Direction { PREVIOUS, NEXT }
+const { PREVIOUS, NEXT } = Direction;
 
 //	Initialize _________________________________________________________________
 
@@ -35,6 +38,8 @@ export class L13DiffListComponent extends L13Element<L13DiffListViewModel> {
 	
 	public disabled:boolean = false;
 	
+	public tabIndex = 0;
+	
 	private cacheSelectionHistory:HTMLElement[] = [];
 	private cacheSelectedListItems:HTMLElement[] = [];
 	private cacheListItemViews:{ [name:string]:HTMLElement } = {};
@@ -45,8 +50,68 @@ export class L13DiffListComponent extends L13Element<L13DiffListViewModel> {
 		
 		super();
 		
+		window.addEventListener('focus', () => {
+			
+			if (this.cacheSelectionHistory.length) this.focus();
+			
+		});
+		
+		this.addEventListener('focus', () => {
+			
+			this.list.classList.add('-focus');
+			
+		});
+		
+		this.addEventListener('blur', () => {
+			
+			this.list.classList.remove('-focus');
+			
+		});
+		
+		this.addEventListener('keydown', (event) => {
+			
+			if (this.disabled) return;
+			
+			switch (event.key) {
+				case 'F12': // Debug Mode
+					if (event.metaKey && event.ctrlKey && event.altKey && event.shiftKey) changePlatform();
+					break;
+				case 'Escape':
+					this.unselect();
+					break;
+				case 'Enter':
+					this.getIdsBySelection().forEach((id) => {
+						
+						vscode.postMessage({
+							command: event.ctrlKey ? 'open:diffToSide' : 'open:diff',
+							diff: this.viewmodel.getDiffById(id),
+						});
+						
+					});
+					break;
+				case 'ArrowUp':
+					this.selectPreviousOrNext(PREVIOUS, event);
+					break;
+				case 'ArrowDown':
+					this.selectPreviousOrNext(NEXT, event);
+					break;
+				case 'PageUp':
+					if (!isMacOs) this.selectPreviousOrNext(PREVIOUS, event);
+					break;
+				case 'PageDown':
+					if (!isMacOs) this.selectPreviousOrNext(NEXT, event);
+				case 'Home':
+					if (!isMacOs) this.selectPreviousOrNext(PREVIOUS, event);
+					break;
+				case 'End':
+					if (!isMacOs) this.selectPreviousOrNext(NEXT, event);
+					break;
+			}
+			
+		});
+		
 		this.list.addEventListener('click', ({ target, metaKey, ctrlKey, shiftKey, offsetX }) => {
-	
+			
 			if (this.disabled) return;
 			
 			if (this.list.firstChild && offsetX > (<HTMLElement>this.list.firstChild).offsetWidth) return;
@@ -56,10 +121,9 @@ export class L13DiffListComponent extends L13Element<L13DiffListViewModel> {
 				return;
 			}
 			
-			const selected = this.list.querySelectorAll('.-selected');
 			const parentNode = <HTMLElement>(<HTMLElement>target).parentNode;
 			
-			if (selected.length) {
+			if (this.cacheSelectionHistory.length) {
 			//	On macOS metaKey overrides shiftKey if both keys are pressed
 				if (isMacOs && shiftKey && !metaKey || !isMacOs && shiftKey) {
 					const lastSelection = this.cacheSelectionHistory[this.cacheSelectionHistory.length - 1];
@@ -72,20 +136,8 @@ export class L13DiffListComponent extends L13Element<L13DiffListViewModel> {
 						//	On Linux always last clicked item will be remembered
 							this.cacheSelectionHistory = [isWindows ? lastSelection : parentNode];
 						}
-						let useSelect = false;
 						if (this.cacheSelectedListItems.length) this.cacheSelectedListItems.forEach((element) => element.classList.remove('-selected'));
-						const elements = this.list.querySelectorAll('l13-diff-list-row');
-						if (elements) {
-							this.cacheSelectedListItems = Array.prototype.slice.call(elements).filter((element) => {
-								
-								if (useSelect || element === parentNode || element === lastSelection) {
-									if (element === parentNode || element === lastSelection) useSelect = !useSelect;
-									element.classList.add('-selected');
-									return true;
-								}
-								
-							});
-						}
+						this.cacheSelectedListItems = this.selectRange(parentNode, lastSelection);
 					} else this.selectListItem(parentNode);
 				} else if (isMetaKey(ctrlKey, metaKey)) {
 					parentNode.classList.toggle('-selected');
@@ -128,21 +180,222 @@ export class L13DiffListComponent extends L13Element<L13DiffListViewModel> {
 	
 	private detectCopy () :void {
 		
-		if (this.list.querySelector('.-selected')) actionsService.model('actions').enableCopy();
+		if (this.cacheSelectionHistory.length) actionsService.model('actions').enableCopy();
 		else actionsService.model('actions').disableCopy();
 		
 	}
 	
-	private selectListItem (parentNode:HTMLElement) {
+	private selectListItem (element:HTMLElement) {
 	
-		parentNode.classList.add('-selected');
-		this.cacheSelectionHistory.push(parentNode);
+		element.classList.add('-selected');
+		
+		this.cacheSelectionHistory.push(element);
 		this.cacheSelectedListItems = [];
+		
 		actionsService.model('actions').enableCopy();
 		
 	}
 	
-	public select (type:string, addToSelection:boolean = false) {
+	private selectRange (from:HTMLElement, to:HTMLElement) {
+		
+		const elements:HTMLElement[] = [];
+		
+		[from, to] = from.offsetTop < to.offsetTop ? [from, to] : [to, from];
+		
+		while (from !== to) {
+			from.classList.add('-selected');
+			elements[elements.length] = from;
+			from = <HTMLElement>from.nextElementSibling;
+		}
+		
+		to.classList.add('-selected');
+		elements[elements.length] = to;
+		
+		return elements;
+		
+	}
+	
+	private selectItem (element:HTMLElement) {
+		
+		element.classList.add('-selected');
+		this.cacheSelectionHistory.push(element);
+		scrollElementIntoView(this, element);
+		
+	}
+	
+	private selectNoneItem (element:HTMLElement, shiftKey:boolean) {
+		
+		if (!shiftKey && this.cacheSelectionHistory.length > 1) {
+			this.unselect();
+			element.classList.add('-selected');
+			this.cacheSelectionHistory.push(element);
+			actionsService.model('actions').enableCopy();
+		}
+		
+		scrollElementIntoView(this, element);
+		
+	}
+	
+	private getFirstItem () {
+		
+		return <HTMLElement>this.list.firstElementChild;
+		
+	}
+	
+	private getLastItem () {
+		
+		return <HTMLElement>this.list.lastElementChild;
+		
+	}
+	
+	private getPreviousPageItem (currentElement:HTMLElement, viewStart:number) {
+		
+		let previousElementSibling:HTMLElement;
+		
+		while ((previousElementSibling = <HTMLElement>currentElement.previousElementSibling)) {
+			if (previousElementSibling.offsetTop > viewStart) {
+				currentElement = previousElementSibling;
+				continue;
+			}
+			break;
+		}
+		
+		return currentElement;
+		
+	}
+	
+	private getNextPageItem (currentElement:HTMLElement, viewEnd:number) {
+		
+		let nextElementSibling:HTMLElement;
+		
+		while ((nextElementSibling = <HTMLElement>currentElement.nextElementSibling)) {
+			if (nextElementSibling.offsetTop + nextElementSibling.offsetHeight < viewEnd) {
+				currentElement = nextElementSibling;
+				continue;
+			}
+			break;
+		}
+		
+		return currentElement;
+		
+	}
+	
+	private selectPreviousOrNextItem (element:HTMLElement, shiftKey:boolean) {
+		
+		if (!shiftKey) this.unselect();
+		
+		this.cacheSelectionHistory.push(element);
+		element.classList.add('-selected');
+		scrollElementIntoView(this, element);
+		actionsService.model('actions').enableCopy();
+		
+	}
+	
+	private selectFirstOrLastItem (from:HTMLElement, to:HTMLElement, shiftKey:boolean) {
+		
+		if (!shiftKey) {
+			this.unselect();
+			to.classList.add('-selected');
+			actionsService.model('actions').enableCopy();
+		} else {
+			if (isWindows) {
+				if (this.cacheSelectedListItems.length) this.cacheSelectedListItems.forEach((element) => element.classList.remove('-selected'));
+				if (this.cacheSelectionHistory.length > 1) {
+					this.cacheSelectionHistory.pop();
+					from = this.cacheSelectionHistory[this.cacheSelectionHistory.length - 1];
+				}
+			}
+			this.cacheSelectedListItems = this.selectRange(from, to);
+		}
+		
+		this.cacheSelectionHistory.push(to);
+		scrollElementIntoView(this, to);
+		
+	}
+	
+	private selectPreviousOrNextPageItem (currentElement:HTMLElement, lastSelection:HTMLElement, shiftKey:boolean) {
+		
+		if (!shiftKey) {
+			this.unselect();
+			currentElement.classList.add('-selected');
+			actionsService.model('actions').enableCopy();
+		} else {
+			this.cacheSelectedListItems = this.selectRange(lastSelection, currentElement);
+		}
+		
+		this.cacheSelectionHistory.push(currentElement);
+		scrollElementIntoView(this, currentElement);
+		
+	}
+	
+	private selectPreviousOrNext (direction:Direction, event:KeyboardEvent) {
+		
+		if (!this.list.firstChild) return;
+		
+		actionsService.model('actions').enableCopy();
+		event.preventDefault();
+		
+		const lastSelection = this.cacheSelectionHistory[this.cacheSelectionHistory.length - 1];
+		
+		if (direction === NEXT) this.selectNext(event, lastSelection);
+		else this.selectPrevious(event, lastSelection);
+		
+	}
+	
+	private selectPrevious ({ altKey, shiftKey, key }:KeyboardEvent, lastSelection:HTMLElement) {
+		
+		if (isMacOs) {
+			if (!lastSelection) this.selectItem(altKey ? this.getFirstItem() : this.getLastItem());
+			else if (!lastSelection.previousElementSibling) this.selectNoneItem(lastSelection, shiftKey);
+			else if (altKey) this.selectFirstOrLastItem(lastSelection, this.getFirstItem(), shiftKey);
+			else this.selectPreviousOrNextItem(<HTMLElement>lastSelection.previousElementSibling, shiftKey);
+		} else {
+			if (key === 'ArrowUp') {
+				if (!lastSelection) this.selectItem(this.getLastItem());
+				else if (!lastSelection.previousElementSibling) this.selectNoneItem(lastSelection, shiftKey);
+				else this.selectPreviousOrNextItem(<HTMLElement>lastSelection.previousElementSibling, shiftKey);
+			} else if (key === 'PageUp') {
+				const viewStart = this.scrollTop - 1; // Why does - 1 fixes the issue???
+				let currentElement = this.getPreviousPageItem(this.getLastItem(), viewStart);
+				if (!lastSelection) this.selectItem(currentElement);
+				if (currentElement === lastSelection) currentElement = this.getPreviousPageItem(lastSelection, viewStart - this.offsetHeight);
+				this.selectPreviousOrNextPageItem(currentElement, lastSelection, shiftKey);
+			} else if (key === 'Home') {
+				if (!lastSelection) this.selectItem(this.getFirstItem());
+				else this.selectFirstOrLastItem(lastSelection, this.getFirstItem(), shiftKey);
+			}
+		}
+		
+	}
+	
+	private selectNext ({ altKey, shiftKey, key }:KeyboardEvent, lastSelection:HTMLElement) {
+		
+		if (isMacOs) {
+			if (!lastSelection) this.selectItem(altKey ? this.getLastItem() : this.getFirstItem());
+			else if (!lastSelection.nextElementSibling) this.selectNoneItem(lastSelection, shiftKey);
+			else if (altKey) this.selectFirstOrLastItem(lastSelection, this.getLastItem(), shiftKey);
+			else this.selectPreviousOrNextItem(<HTMLElement>lastSelection.nextElementSibling, shiftKey);
+		} else {
+			if (key === 'ArrowDown') {
+				if (!lastSelection) this.selectItem(this.getFirstItem());
+				else if (!lastSelection.nextElementSibling) this.selectNoneItem(lastSelection, shiftKey);
+				else this.selectPreviousOrNextItem(<HTMLElement>lastSelection.nextElementSibling, shiftKey);
+			} else if (key === 'PageDown') {
+				const viewHeight = this.offsetHeight;
+				const viewEnd = this.scrollTop + viewHeight + 1; // Why does + 1 fixes the issue???
+				let currentElement = this.getNextPageItem(this.getFirstItem(), viewEnd);
+				if (!lastSelection) this.selectItem(currentElement);
+				if (currentElement === lastSelection) currentElement = this.getNextPageItem(lastSelection, viewEnd + viewHeight);
+				this.selectPreviousOrNextPageItem(currentElement, lastSelection, shiftKey);
+			} else if (key === 'End') {
+				if (!lastSelection) this.selectItem(this.getLastItem());
+				else this.selectFirstOrLastItem(lastSelection, this.getLastItem(), shiftKey);
+			}
+		}
+		
+	}
+	
+	public selectByStatus (type:string, addToSelection:boolean = false) {
 		
 		if (!addToSelection) this.unselect();
 		
@@ -160,9 +413,9 @@ export class L13DiffListComponent extends L13Element<L13DiffListViewModel> {
 		
 		const elements = this.list.querySelectorAll('.-selected');
 	
-		if (elements) elements.forEach((element) => element.classList.remove('-selected'));
+		if (elements.length) elements.forEach((element) => element.classList.remove('-selected'));
 		
-		this.detectCopy();
+		actionsService.model('actions').disableCopy();
 		
 	}