+++ src/views/main.ts
@@ -14,10 +14,6 @@ import './components/l13-diff-compare/l13-diff-compare.component';
 import './components/l13-diff-compare/l13-diff-compare.service';
 import './components/l13-diff-compare/l13-diff-compare.viewmodel';
 
-import './components/l13-diff-folders/l13-diff-folders.component';
-import './components/l13-diff-folders/l13-diff-folders.service';
-import './components/l13-diff-folders/l13-diff-folders.viewmodel';
-
 import './components/l13-diff-input/l13-diff-input.component';
 import './components/l13-diff-input/l13-diff-input.service';
 import './components/l13-diff-input/l13-diff-input.viewmodel';
@@ -37,10 +33,6 @@ import './components/l13-diff-panel/l13-diff-panel.viewmodel';
 import './components/l13-diff-swap/l13-diff-swap.component';
 import './components/l13-diff-swap/l13-diff-swap.service';
 import './components/l13-diff-swap/l13-diff-swap.viewmodel';
-
-import './components/l13-diff-tools/l13-diff-tools.component';
-import './components/l13-diff-tools/l13-diff-tools.service';
-import './components/l13-diff-tools/l13-diff-tools.viewmodel';
 
 import './components/l13-diff-views/l13-diff-views.component';
 import './components/l13-diff-views/l13-diff-views.service';