+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Change Log
 All notable changes to the "L13 Diff" extension will be documented in this file.
 
-## [0.13.0] - 2018-05-12
+## [0.14.0] - 2019-05-26
+
+### Added
+- Added keyboard support for selection in list view. Please see README.md
+
+### Fixed
+- Fixed a bug if an item was select in the menu with enter, but value was not set.
+- Fixed async loading bug if VS Code for Windows is running on a virtual machine.
+- Fixed invisible selection in input fields for VS Codes default themes.
+- Fixed invisible checkbox icon for light themes.
+- Fixed scroll into view bug for menu component.
+- Fixed a bug if files are selected but copy buttons were still disabled.
+- Fixed wrong year in changelog.
+
+## [0.13.0] - 2019-05-12
 - Initial release